refactor: extract calendar search UI
