+++ app/src/main/java/com/kgs/calendar/ui/KgsCalendarApp.kt
@@ -2893,731 +2893,6 @@ private fun TaskDrawer(
 }
 
 @Composable
-private fun CalendarSearchOverlay(
-    visible: Boolean,
-    query: String,
-    results: List<EventEntity>,
-    taskResults: List<TaskEntity>,
-    allTasksForHierarchy: List<TaskEntity>,
-    taskColorMode: TaskColorMode,
-    subtasksExpandedByDefault: Boolean,
-    onQueryChange: (String) -> Unit,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onEventClick: (EventEntity) -> Unit,
-    onTaskClick: (TaskEntity) -> Unit,
-    onClose: () -> Unit,
-) {
-    AnimatedVisibility(
-        visible = visible,
-        enter = slideInVertically(animationSpec = tween(MotionLong, easing = MotionEmphasized)) { -it / 5 } +
-            scaleIn(initialScale = 0.98f, animationSpec = tween(MotionLong, easing = MotionEmphasized)) +
-            fadeIn(animationSpec = tween(MotionMedium, delayMillis = 50, easing = MotionStandard)),
-        exit = slideOutVertically(animationSpec = tween(MotionMedium, easing = MotionStandardAccelerate)) { -it / 6 } +
-            scaleOut(targetScale = 0.98f, animationSpec = tween(MotionMedium, easing = MotionStandardAccelerate)) +
-            fadeOut(animationSpec = tween(MotionShort, easing = MotionStandardAccelerate)),
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.surface),
-        ) {
-            SearchTopBar(query = query, onQueryChange = onQueryChange, onClose = onClose)
-            SearchResultsList(
-                query = query,
-                eventResults = results,
-                taskResults = taskResults,
-                allTasksForHierarchy = allTasksForHierarchy,
-                taskColorMode = taskColorMode,
-                subtasksExpandedByDefault = subtasksExpandedByDefault,
-                onEventClick = onEventClick,
-                onTaskClick = onTaskClick,
-                onTaskStatusChanged = onTaskStatusChanged,
-            )
-        }
-    }
-}
-
-@Composable
-private fun SearchTopBar(query: String, onQueryChange: (String) -> Unit, onClose: () -> Unit) {
-    val focusRequester = remember { FocusRequester() }
-    var focused by remember { mutableStateOf(false) }
-    val statusTop = WindowInsets.statusBars.asPaddingValues().calculateTopPadding()
-    LaunchedEffect(Unit) {
-        focusRequester.requestFocus()
-    }
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(statusTop + 62.dp)
-            .background(MaterialTheme.colorScheme.background)
-            .padding(start = 16.dp, end = 14.dp, top = statusTop),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        IconButton(onClick = onClose, modifier = Modifier.size(38.dp)) {
-            Icon(Icons.Default.ArrowBack, contentDescription = stringResource(R.string.back), tint = WarmInk, modifier = Modifier.size(25.dp))
-        }
-        BasicTextField(
-            value = query,
-            onValueChange = onQueryChange,
-            singleLine = true,
-            textStyle = MaterialTheme.typography.headlineSmall.copy(
-                color = WarmInk,
-                fontSize = 22.sp,
-                lineHeight = 26.sp,
-                fontWeight = FontWeight.Normal,
-            ),
-            modifier = Modifier
-                .weight(1f)
-                .fillMaxWidth()
-                .pointerInput(Unit) {
-                    detectTapGestures {
-                        focused = true
-                        focusRequester.requestFocus()
-                    }
-                }
-                .focusRequester(focusRequester)
-                .onFocusChanged { focused = it.isFocused },
-            cursorBrush = SolidColor(WarmBrown),
-            decorationBox = { innerTextField ->
-                Box(contentAlignment = Alignment.CenterStart) {
-                    if (query.isBlank() && !focused) {
-                        Text(stringResource(R.string.search), color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 22.sp, lineHeight = 26.sp)
-                    }
-                    innerTextField()
-                }
-            },
-        )
-        IconButton(
-            modifier = Modifier.size(38.dp),
-            onClick = {
-            if (query.isBlank()) onClose() else onQueryChange("")
-        }) {
-            Icon(Icons.Default.Close, contentDescription = stringResource(R.string.close), tint = WarmInk, modifier = Modifier.size(25.dp))
-        }
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun SearchResultsList(
-    query: String,
-    eventResults: List<EventEntity>,
-    taskResults: List<TaskEntity>,
-    allTasksForHierarchy: List<TaskEntity> = taskResults,
-    taskColorMode: TaskColorMode,
-    subtasksExpandedByDefault: Boolean,
-    onEventClick: (EventEntity) -> Unit,
-    onTaskClick: (TaskEntity) -> Unit,
-    onTaskStatusChanged: (String, String) -> Unit,
-    showSearchIntro: Boolean = true,
-    autoScrollToNow: Boolean = query.isNotBlank() || !showSearchIntro,
-    emptyMessage: String? = null,
-    scrollRequestKey: Int = 0,
-    scrollTargetDate: LocalDate? = null,
-    stickyHeaderBackground: Color = MaterialTheme.colorScheme.surface,
-    showTaskChains: Boolean = true,
-    expandMultiDayEventSpans: Boolean = false,
-) {
-    val resolvedEmptyMessage = emptyMessage ?: stringResource(R.string.no_results)
-    val noDateLabel = stringResource(R.string.no_date)
-    val hierarchySubset = remember(query, showSearchIntro, showTaskChains, taskResults, allTasksForHierarchy) {
-        if (!showTaskChains) {
-            TaskHierarchySubset(emptyList())
-        } else {
-            when {
-                query.isNotBlank() -> allTasksForHierarchy.searchHierarchySubsetForMatches(taskResults)
-                !showSearchIntro -> TaskHierarchySubset(
-                    tasks = allTasksForHierarchy.chainSubsetForTargets(
-                        targets = taskResults,
-                        includeTargetDescendants = true,
-                    ),
-                )
-                else -> TaskHierarchySubset(tasks = taskResults)
-            }
-        }
-    }
-    val hierarchyExpandedByDefault = when {
-        query.isNotBlank() -> false
-        !showSearchIntro -> true
-        else -> subtasksExpandedByDefault
-    }
-    val taskHierarchy = rememberTaskHierarchyPresentation(
-        tasks = hierarchySubset.tasks,
-        expandedByDefault = hierarchyExpandedByDefault,
-        defaultExpandedResourceHrefs = hierarchySubset.defaultExpandedResourceHrefs,
-    )
-    val taskGroups = remember(taskHierarchy.entries) {
-        buildList {
-            var current = mutableListOf<TaskHierarchyEntry>()
-            taskHierarchy.entries.forEach { entry ->
-                if (entry.depth == 0 && current.isNotEmpty()) {
-                    add(CalendarSearchResult.TaskGroup(current.toList()))
-                    current = mutableListOf()
-                }
-                current += entry
-            }
-            if (current.isNotEmpty()) add(CalendarSearchResult.TaskGroup(current.toList()))
-        }
-    }
-    val flatTaskItems = remember(showTaskChains, taskResults) {
-        if (showTaskChains) {
-            emptyList()
-        } else {
-            taskResults
-                .distinctBy { it.resourceHref }
-                .filter { it.agendaSortMillis() != null }
-                .map { CalendarSearchResult.TaskItem(it) }
-        }
-    }
-    val eventItems = remember(eventResults, taskResults, expandMultiDayEventSpans) {
-        if (expandMultiDayEventSpans) {
-            buildAgendaEventResults(eventResults, taskResults)
-        } else {
-            eventResults.map { CalendarSearchResult.Event(it) }
-        }
-    }
-    val results = remember(eventItems, taskGroups, flatTaskItems) {
-        (eventItems + taskGroups + flatTaskItems)
-            .sortedWith(
-                compareBy<CalendarSearchResult> { it.sortMillis == null }
-                    .thenBy { it.sortMillis ?: Long.MAX_VALUE },
-            )
-    }
-    val groupedResults = remember(results) {
-        results.groupBy { item -> item.date?.year?.toString() ?: noDateLabel }
-    }
-    val today = LocalCalendarTimeSnapshot.current.today
-    val listState = rememberLazyListState()
-    val density = LocalDensity.current
-    val firstFutureKey = remember(results, today) {
-        results.firstOrNull { it.isFutureSearchSection(today) }?.stableKey()
-    }
-    var initialAutoScrollHandled by remember { mutableStateOf(false) }
-    var lastHandledScrollRequestKey by remember { mutableStateOf(scrollRequestKey) }
-    val firstFutureLazyIndex = remember(groupedResults, firstFutureKey, today) {
-        if (firstFutureKey == null) null else {
-            var lazyIndex = if (showSearchIntro && query.isNotBlank()) 1 else 0
-            var found: Int? = null
-            var dividerInserted = false
-            for (groupItems in groupedResults.values) {
-                if (found != null) break
-                lazyIndex++ // sticky year header
-                for (item in groupItems) {
-                    if (!dividerInserted && item.isFutureSearchSection(today)) {
-                        lazyIndex++ // past/future divider
-                        dividerInserted = true
-                    }
-                    if (item.stableKey() == firstFutureKey) {
-                        found = lazyIndex
-                        break
-                    }
-                    lazyIndex++
-                }
-            }
-            found
-        }
-    }
-    val requestedDateLazyIndex = remember(groupedResults, scrollTargetDate, showSearchIntro, query, today) {
-        val targetDate = scrollTargetDate ?: return@remember null
-        var lazyIndex = if (showSearchIntro && query.isNotBlank()) 1 else 0
-        var found: Int? = null
-        var dividerInserted = false
-        for (groupItems in groupedResults.values) {
-            if (found != null) break
-            lazyIndex++ // sticky year header
-            for (item in groupItems) {
-                if (!dividerInserted && item.isFutureSearchSection(today)) {
-                    lazyIndex++ // past/future divider
-                    dividerInserted = true
-                }
-                val itemDate = item.date
-                if (itemDate != null && !itemDate.isBefore(targetDate)) {
-                    found = lazyIndex
-                    break
-                }
-                lazyIndex++
-            }
-        }
-        found
-    }
-    LaunchedEffect(query, firstFutureLazyIndex, requestedDateLazyIndex, scrollRequestKey) {
-        if (autoScrollToNow) {
-            val explicitAgendaRequest = !showSearchIntro && scrollRequestKey != lastHandledScrollRequestKey
-            val requestedIndex = if (explicitAgendaRequest && requestedDateLazyIndex != null) requestedDateLazyIndex else null
-            val baseTarget = requestedIndex ?: firstFutureLazyIndex
-            if (baseTarget == null) return@LaunchedEffect
-            val targetIndex = if (showSearchIntro) {
-                (baseTarget - 2).coerceAtLeast(0)
-            } else {
-                (baseTarget - 2).coerceAtLeast(0)
-            }
-            val targetOffset = if (showSearchIntro) 0 else -with(density) { 38.dp.roundToPx() }
-            if (initialAutoScrollHandled && explicitAgendaRequest) {
-                listState.animateScrollToItem(targetIndex, targetOffset)
-            } else {
-                listState.scrollToItem(targetIndex, targetOffset)
-            }
-            initialAutoScrollHandled = true
-            lastHandledScrollRequestKey = scrollRequestKey
-        }
-    }
-    LazyColumn(
-        state = listState,
-        modifier = Modifier
-            .fillMaxSize(),
-        contentPadding = PaddingValues(horizontal = 18.dp, vertical = 20.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        if (query.isBlank() && showSearchIntro) {
-            item {
-                Text(stringResource(R.string.search_events_tasks), color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 14.sp, lineHeight = 18.sp, modifier = Modifier.padding(start = 62.dp, top = 18.dp))
-            }
-        } else {
-            if (showSearchIntro && query.isNotBlank()) {
-                item {
-                    Text(
-                        stringResource(R.string.search_may_miss_old_results),
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        fontSize = 14.sp,
-                        lineHeight = 18.sp,
-                        modifier = Modifier.padding(start = 62.dp, bottom = 8.dp),
-                    )
-                }
-            }
-            if (results.isEmpty()) {
-                item {
-                    Text(resolvedEmptyMessage, color = WarmInk, fontSize = 16.sp, lineHeight = 20.sp, modifier = Modifier.padding(start = 62.dp, top = 18.dp))
-                }
-            } else {
-                var futureDividerInserted = false
-                groupedResults.forEach { (group, groupItems) ->
-                    stickyHeader(key = "search-year-$group") {
-                        Box(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .background(stickyHeaderBackground.copy(alpha = 0.97f))
-                                .padding(start = 62.dp, top = 7.dp, bottom = 7.dp),
-                        ) {
-                            Text(
-                                group,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                fontSize = 15.sp,
-                                lineHeight = 19.sp,
-                                fontWeight = FontWeight.SemiBold,
-                            )
-                        }
-                    }
-                    var lastDateKey: String? = null
-                    groupItems.forEach { item ->
-                        val dateKey = item.date?.toString() ?: "__no_date__"
-                        val showDate = dateKey != lastDateKey
-                        if (!futureDividerInserted && item.isFutureSearchSection(today)) {
-                            item(key = "search-past-future-divider") {
-                                SearchPastFutureDivider()
-                            }
-                            futureDividerInserted = true
-                        }
-                        item(key = item.stableKey()) {
-                        when (item) {
-                            is CalendarSearchResult.Event -> SearchResultCard(
-                                event = item.event,
-                                displayDate = item.date,
-                                spanEndDate = item.spanEndDate,
-                                showDate = showDate,
-                                onClick = { onEventClick(item.event) },
-                            )
-                            is CalendarSearchResult.TaskItem -> SearchTaskResultCard(
-                                task = item.task,
-                                showDate = showDate,
-                                taskColorMode = taskColorMode,
-                                onTaskStatusChanged = onTaskStatusChanged,
-                                hierarchyDepth = 0,
-                                hierarchyContinuationLevels = emptySet(),
-                                hierarchyLastSibling = true,
-                                hasSubtasks = false,
-                                subtasksExpanded = false,
-                                onToggleSubtasks = {},
-                                onClick = { onTaskClick(item.task) },
-                            )
-                            is CalendarSearchResult.TaskGroup -> Column(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .graphicsLayer { clip = false },
-                            ) {
-                                item.entries.forEach { entry ->
-                                    AnimatedTaskHierarchyEntry(entry) {
-                                        SearchTaskResultCard(
-                                            task = entry.task,
-                                            showDate = showDate,
-                                            taskColorMode = taskColorMode,
-                                            onTaskStatusChanged = onTaskStatusChanged,
-                                            hierarchyDepth = entry.depth,
-                                            hierarchyContinuationLevels = entry.continuationLevels,
-                                            hierarchyLastSibling = entry.lastSibling,
-                                            hasSubtasks = entry.hasChildren,
-                                            subtasksExpanded = entry.expanded,
-                                            onToggleSubtasks = { taskHierarchy.toggle(entry.task) },
-                                            onClick = { onTaskClick(entry.task) },
-                                        )
-                                    }
-                                }
-                            }
-                        }
-                        }
-                        lastDateKey = dateKey
-                    }
-                }
-                item { Spacer(Modifier.height(80.dp)) }
-            }
-        }
-    }
-}
-
-private sealed interface CalendarSearchResult {
-    val sortMillis: Long?
-    val date: LocalDate?
-
-    data class Event(
-        val event: EventEntity,
-        override val date: LocalDate = event.startsAtMillis.toDate(),
-        val spanEndDate: LocalDate = date,
-    ) : CalendarSearchResult {
-        override val sortMillis: Long = if (date == event.startsAtMillis.toDate()) {
-            event.startsAtMillis
-        } else {
-            date.startOfDayMillis()
-        }
-    }
-
-    data class TaskGroup(val entries: List<TaskHierarchyEntry>) : CalendarSearchResult {
-        private val root: TaskEntity = entries.first().task
-        override val sortMillis: Long? = entries
-            .mapNotNull { it.task.startAtMillis ?: it.task.dueAtMillis }
-            .minOrNull()
-            ?: root.startAtMillis
-            ?: root.dueAtMillis
-        override val date: LocalDate? = sortMillis?.toDate()
-    }
-
-    data class TaskItem(val task: TaskEntity) : CalendarSearchResult {
-        override val sortMillis: Long? = task.agendaSortMillis()
-        override val date: LocalDate? = sortMillis?.toDate()
-    }
-}
-
-private fun CalendarSearchResult.stableKey(): String = when (this) {
-    is CalendarSearchResult.Event -> "event-${event.resourceHref}-${event.startsAtMillis}-$date-$spanEndDate"
-    is CalendarSearchResult.TaskGroup -> "task-group-${entries.first().task.resourceHref}"
-    is CalendarSearchResult.TaskItem -> "task-${task.resourceHref}-${task.agendaSortMillis() ?: 0L}"
-}
-
-private fun CalendarSearchResult.isFutureSearchSection(today: LocalDate): Boolean =
-    date?.isBefore(today) != true
-
-private fun buildAgendaEventResults(
-    events: List<EventEntity>,
-    tasks: List<TaskEntity>,
-): List<CalendarSearchResult.Event> {
-    val taskDates = tasks
-        .flatMap { it.visibleDates() }
-        .toSet()
-    val eventDatesByResource = events.associate { event ->
-        event.resourceHref to event.visibleAgendaDates()
-    }
-    return events.flatMap { event ->
-        val dates = eventDatesByResource[event.resourceHref].orEmpty()
-        if (dates.size <= 1) {
-            listOf(CalendarSearchResult.Event(event, dates.firstOrNull() ?: event.startsAtMillis.toDate()))
-        } else {
-            val interruptionDates = dates.filterTo(mutableSetOf()) { date ->
-                date in taskDates || events.any { other ->
-                    other.resourceHref != event.resourceHref && eventDatesByResource[other.resourceHref].orEmpty().contains(date)
-                }
-            }
-            event.toAgendaSpanResults(dates, interruptionDates)
-        }
-    }
-}
-
-private fun EventEntity.toAgendaSpanResults(
-    dates: List<LocalDate>,
-    interruptionDates: Set<LocalDate>,
-): List<CalendarSearchResult.Event> {
-    val sortedDates = dates.sorted()
-    if (sortedDates.isEmpty()) return emptyList()
-    val results = mutableListOf<CalendarSearchResult.Event>()
-    var segmentStart: LocalDate? = null
-    var previous: LocalDate? = null
-
-    fun flushSegment(end: LocalDate) {
-        val start = segmentStart ?: return
-        if (!end.isBefore(start)) {
-            results += CalendarSearchResult.Event(this, start, end)
-        }
-        segmentStart = null
-    }
-
-    sortedDates.forEach { date ->
-        val last = previous
-        if (last != null && last.plusDays(1) != date) {
-            flushSegment(last)
-        }
-        if (date in interruptionDates) {
-            flushSegment(date.minusDays(1))
-            results += CalendarSearchResult.Event(this, date, date)
-        } else if (segmentStart == null) {
-            segmentStart = date
-        }
-        previous = date
-    }
-    previous?.let(::flushSegment)
-    return results
-}
-
-@Composable
-private fun SearchPastFutureDivider() {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .horizontalBleed(18.dp)
-            .padding(top = 12.dp, bottom = 12.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(5.dp),
-    ) {
-        Text(
-            stringResource(R.string.past),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            fontSize = 11.sp,
-            lineHeight = 13.sp,
-            fontWeight = FontWeight.SemiBold,
-        )
-        HorizontalDivider(color = WarmLine)
-        Text(
-            stringResource(R.string.future),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            fontSize = 11.sp,
-            lineHeight = 13.sp,
-            fontWeight = FontWeight.SemiBold,
-        )
-    }
-}
-
-@Composable
-private fun SearchResultCard(
-    event: EventEntity,
-    showDate: Boolean,
-    onClick: () -> Unit,
-    displayDate: LocalDate = event.startsAtMillis.toDate(),
-    spanEndDate: LocalDate = displayDate,
-) {
-    val isPast = event.endsAtMillis < System.currentTimeMillis()
-    val isSpan = spanEndDate.isAfter(displayDate)
-    val attendees = remember(event.attendeesJson) { event.attendeesJson.toCalendarParticipants() }
-    val rowHeight = when {
-        isSpan -> AgendaMultiDaySpanCardHeight
-        attendees.isNotEmpty() -> 76.dp
-        else -> 64.dp
-    }
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-        verticalAlignment = Alignment.Top,
-    ) {
-        val visuals = event.cardVisuals(muted = isPast, darkPalette = CurrentDarkPalette)
-        SearchDateColumn(
-            date = displayDate,
-            muted = isPast,
-            visible = showDate,
-            endDate = spanEndDate.takeIf { isSpan },
-            height = if (isSpan) rowHeight else null,
-        )
-        val eventTextStyle = tentativeReadableTextStyle(event.isTentative() && !isPast)
-        val shape = RoundedCornerShape(13.dp)
-        val pendingAlpha = pendingDeleteAlpha(event.resourceHref)
-        Box(
-            modifier = Modifier
-                .weight(1f)
-                .height(rowHeight)
-                .alpha(pendingAlpha)
-                .clip(shape)
-                .background(visuals.background)
-                .then(
-                    when {
-                        visuals.dashedBorder && visuals.borderColor != null -> Modifier.dashedBorder(visuals.borderColor, 13.dp)
-                        visuals.borderColor != null -> Modifier.border(1.dp, visuals.borderColor, shape)
-                        else -> Modifier
-                    },
-                )
-                .clickable(onClick = onClick)
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .bottomEdgeFadeMask(10.dp)
-                    .padding(start = 12.dp, top = 7.dp, end = 12.dp, bottom = 0.dp)
-                    .padding(end = if (attendees.isNotEmpty()) 54.dp else 0.dp),
-                verticalArrangement = Arrangement.Top,
-            ) {
-                FadingTimedText(event.title, color = visuals.contentColor, fontSize = 14.sp, lineHeight = 17.sp, fontWeight = FontWeight.SemiBold, maxLines = 1, textDecoration = visuals.textDecoration, shadow = eventTextStyle.shadow)
-                FadingTimedText(
-                    if (isSpan) event.localizedAgendaSpanLabel(displayDate, spanEndDate) else event.localizedTimeLabel(),
-                    color = visuals.contentColor.copy(alpha = if (isPast) 0.72f else 0.92f),
-                    fontSize = 12.sp,
-                    lineHeight = 15.sp,
-                    maxLines = 1,
-                    textDecoration = visuals.textDecoration,
-                    shadow = eventTextStyle.shadow,
-                )
-                event.location?.takeIf { it.isNotBlank() }?.let {
-                    FadingTimedText(it.cardLocationText(event.locationMapVerified), color = visuals.contentColor.copy(alpha = if (isPast) 0.66f else 0.86f), fontSize = 11.sp, lineHeight = 13.sp, maxLines = 1, textDecoration = visuals.textDecoration, shadow = eventTextStyle.shadow)
-                }
-            }
-            EventParticipantStack(
-                attendees = attendees,
-                eventColor = visuals.baseColor,
-                contentColor = visuals.contentColor,
-                modifier = Modifier.align(Alignment.BottomEnd),
-                circleSize = 24.dp,
-                maxVisible = 5,
-                muted = isPast,
-            )
-            PendingMutationBadge(
-                resourceHref = event.resourceHref,
-                modifier = Modifier.align(Alignment.TopEnd).offset(x = 2.dp, y = (-2).dp),
-            )
-        }
-    }
-}
-
-@Composable
-private fun SearchTaskResultCard(
-    task: TaskEntity,
-    showDate: Boolean,
-    taskColorMode: TaskColorMode,
-    onTaskStatusChanged: (String, String) -> Unit,
-    hierarchyDepth: Int = 0,
-    hierarchyContinuationLevels: Set<Int> = emptySet(),
-    hierarchyLastSibling: Boolean = true,
-    hasSubtasks: Boolean = false,
-    subtasksExpanded: Boolean = true,
-    onToggleSubtasks: (() -> Unit)? = null,
-    onClick: () -> Unit,
-) {
-    val taskDate = (task.startAtMillis ?: task.dueAtMillis)?.toDate()
-    val isMuted = task.isInactive()
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .zIndex(taskHierarchyLayerZ(hierarchyDepth) + taskPriorityLayerZ(task, PriorityAnimationsEnabled))
-            .graphicsLayer { clip = false },
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-        verticalAlignment = Alignment.Top,
-    ) {
-        if (hierarchyDepth == 0) {
-            SearchDateColumn(date = taskDate, muted = isMuted, visible = showDate)
-        } else {
-            Spacer(Modifier.width(50.dp))
-        }
-        TaskRow(
-            task = task,
-            taskColorMode = taskColorMode,
-            onTaskStatusChanged = onTaskStatusChanged,
-            prominent = true,
-            hierarchyDepth = hierarchyDepth,
-            hierarchyContinuationLevels = hierarchyContinuationLevels,
-            hierarchyLastSibling = hierarchyLastSibling,
-            hasSubtasks = hasSubtasks,
-            subtasksExpanded = subtasksExpanded,
-            onToggleSubtasks = onToggleSubtasks,
-            modifier = Modifier.weight(1f),
-            outerHorizontalPadding = 0.dp,
-            outerVerticalPadding = 3.dp,
-            connectorStemInset = TaskHierarchyStemInset,
-            priorityMotionEnabled = !isMuted,
-            onClick = onClick,
-        )
-    }
-}
-
-private val AgendaMultiDaySpanCardHeight = 96.dp
-
-@Composable
-private fun SearchDateColumn(
-    date: LocalDate?,
-    muted: Boolean,
-    visible: Boolean = true,
-    endDate: LocalDate? = null,
-    height: Dp? = null,
-) {
-    Column(
-        modifier = Modifier
-            .width(50.dp)
-            .then(if (height != null) Modifier.height(height) else Modifier),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        if (visible) {
-            if (date == null) {
-                Text(
-                    appString(R.string.none),
-                    color = WarmInk.copy(alpha = if (muted) 0.58f else 1f),
-                    fontSize = 10.sp,
-                    lineHeight = 12.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    appString(R.string.date),
-                    color = WarmInk.copy(alpha = if (muted) 0.58f else 1f),
-                    fontSize = 13.sp,
-                    lineHeight = 16.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-            } else {
-                Text(
-                    date.format(DateTimeFormatter.ofPattern("MMM", LocalAppLocale.current)).replace(".", ""),
-                    color = WarmInk.copy(alpha = if (muted) 0.58f else 1f),
-                    fontSize = 11.sp,
-                    lineHeight = 13.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    date.dayOfMonth.toString(),
-                    color = WarmInk.copy(alpha = if (muted) 0.58f else 1f),
-                    fontSize = 20.sp,
-                    lineHeight = 23.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                if (endDate != null) {
-                    Spacer(Modifier.height(3.dp))
-                    Box(
-                        modifier = Modifier
-                            .width(2.dp)
-                            .weight(1f)
-                            .clip(RoundedCornerShape(999.dp))
-                            .background(WarmInk.copy(alpha = if (muted) 0.22f else 0.42f)),
-                    )
-                    Spacer(Modifier.height(3.dp))
-                    Text(
-                        endDate.format(DateTimeFormatter.ofPattern("MMM", LocalAppLocale.current)).replace(".", ""),
-                        color = WarmInk.copy(alpha = if (muted) 0.58f else 1f),
-                        fontSize = 10.sp,
-                        lineHeight = 12.sp,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                    Text(
-                        endDate.dayOfMonth.toString(),
-                        color = WarmInk.copy(alpha = if (muted) 0.58f else 1f),
-                        fontSize = 17.sp,
-                        lineHeight = 19.sp,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
 internal fun MonthView(
     state: CalendarUiState,
     onMonthChanged: (YearMonth) -> Unit,
@@ -6511,9 +5786,9 @@ internal data class TaskHierarchyPresentation(
 internal val TaskHierarchyStemInset = 10.dp
 private val TaskHierarchyLineOverlap = 4.dp
 
-private fun taskHierarchyLayerZ(depth: Int): Float = 100f - depth.coerceAtLeast(0)
+internal fun taskHierarchyLayerZ(depth: Int): Float = 100f - depth.coerceAtLeast(0)
 
-private fun taskPriorityLayerZ(task: TaskEntity, animationsEnabled: Boolean): Float {
+internal fun taskPriorityLayerZ(task: TaskEntity, animationsEnabled: Boolean): Float {
     if (!animationsEnabled || task.isInactive()) return 0f
     val intensity = taskPriorityIntensity(task.priority)
     return if (intensity > 0f) 1_000f + intensity * 100f else 0f
@@ -12251,7 +11526,7 @@ private fun LocalTime.defaultDraftEnd(durationMinutes: Int = 60): LocalTime =
 
 internal fun LocalTime.minuteOfDay(): Int = hour * 60 + minute
 
-private fun LocalDate.startOfDayMillis(): Long =
+internal fun LocalDate.startOfDayMillis(): Long =
     atStartOfDay(ZoneId.systemDefault()).toInstant().toEpochMilli()
 
 internal fun eventDateTimeRangeInvalid(
@@ -12824,13 +12099,13 @@ internal fun buildAllDayOverlayItems(
 }
 
 @Composable
-private fun EventEntity.localizedTimeLabel(): String {
+internal fun EventEntity.localizedTimeLabel(): String {
     if (allDay) return stringResource(R.string.all_day)
     return "${startsAtMillis.toTimeText()} - ${endsAtMillis.toTimeText()}"
 }
 
 @Composable
-private fun EventEntity.localizedAgendaSpanLabel(startDate: LocalDate, endDate: LocalDate): String {
+internal fun EventEntity.localizedAgendaSpanLabel(startDate: LocalDate, endDate: LocalDate): String {
     val dateFormatter = DateTimeFormatter.ofPattern("d. MMM yyyy", LocalAppLocale.current)
     val allDayLabel = appString(R.string.all_day)
     return if (allDay) {
